[refactor] package rename
